Update environment.yml
I have tried to add the packages needed for the PWSQC method. I have duplicated the style of the other R-packages by prefixing "- r-", although I'm not sure that is correct. 
The required R-packages from CRAN are: 
data.table (for preprocessing in PWSQC_0.R)
geosphere (for PWSQC_1)
sf (for PWSQC_1_alternativeversion.R)
s1 (for PWSQC_1_alternativeversion.R)
TeachingDemos (for visualization in PWSQC_4.R)