Added WIP notebook for plotting OpenMRG dataset. Not working yet.
For some reason the time series do not show up when I click a point of a
CML on the map. The plotting code works with the pycomlink example data,
that is used in the current state of the notebook. Hence, the plotting
code is fine in general. Most likely there is some part of the code that
requires a `channel_id` to be presetn, as in the pycomlink example data.
I cannot figure out where, though...