Solution to work with 2 GB of RAM on binder
I now write the data slice by slice to a chunked zarr store and work
from there. I actually developed it on a running binder instance and it
worked and just copied it to my local notebook. Hence, it should work
now after this commit, hopefully...

I also cleand up the notebook and polisehd the text a bit.

Some more work needs to be done, though. I really want to get the map
tiles in the final version, but it is not clear how much work it will be
to switch to using `hv.points` instead of `hv.scatter`.