OpenMRG correct details in metadata
opensense_data_downloader_and_transformer: added comment indicating how file was created, frequency to MHz and length to meter. Add attributes after transform so that they are not removed by the transform. Removed specification of "missing_value" in all attributes.

create_small_openMRG_example: updated to add information on authors of the stored file and further info on resampling and interpolation of gaps. removed encoding specifications so that xarray defaults to normal behaviour.

openMRG_use_case: updated to run with new nc file. transform frequency to GHz and length to km to make it work with pycomlink.