fix `sublink` kwarg in pynncml_basic_example.ipynb
There were changes in the data transformation code, which now only returns a xarray.Dataset with reshaped data and updated dimension (cml_id, sublink_id, time) if it is called with the option to reshape the data. Since reshaping the data takes too long to run (an might crash on binder due to limited RAM) the simplest fix is to not reshape and just use the `sublink` dimension as it is called in the existing OpenMRG NetCDF file.

At some point, the pynncml example could use the smaller example NetCDF from the OpenMRG dataset that is currently part of the sandbox repository.